refactor: split plugin-codegen index.ts into focused modules
Breaks the 1,055-line monolith into six focused files (types.ts,
errors.ts, options.ts, naming.ts, generators.ts, plugin.ts) with a
barrel index.ts, following the plugin-backfill pattern. No public API
changes; all 18 tests pass.

Entire-Checkpoint: a13948e97dff
